Change internal resolution BaseType from float to int
These represent integer pixel sizes.

+++ NAS2D/Renderer/Renderer.h
@@ -169,7 +169,7 @@ protected:
 
 	void driverName(const std::string& name);
 
-	Vector<float> mResolution{1600,900}; /**< Screen resolution. Reasonable default in 2019*/
+	Vector<int> mResolution{1600,900}; /**< Screen resolution. Reasonable default in 2019*/
 
 private:
 	enum class FadeType
